refactor: separate processing and frontend logic
EncryptFile and DecryptFile functions now take an io.WriteCloser for progress instead of a progress bar directly.
They also now accept os.File as an input rather than the file path. This is a part of refactoring decisions made.
This greatly decreases the coupling between the frontend and backend and also increses core's universality.